refactor: move ProjectIdentityContext and RuntimeContext out of _core/
These dataclasses imported upper-layer types (CppAnalyzerConfig,
ProjectIdentity, _indexing.*) violating the inward-dependency rule.
Moved to a new _contexts/ package — a neutral layer that can import
from any direction without creating cycles.

_core/ now contains only true cross-cutting primitives:
- diagnostics, concurrency_context, cancellation_coordinator,
  file_scanner, file_utils, libclang_setup

Updated imports in project_context.py and compilation_environment.py.
